+++ changelog.md
@@ -3,11 +3,19 @@
 All notable changes to this project will be documented in this file.
 For any suggestions/emojis to add, go here: [Discord Link](https://discord.gg/FeYvZu5x)
 
+
+## {Unreleased}
+
+- send different link when texture pack github latest release is updated
+
+
 ## [0.1.0]
+
+Emojifies emojis in the party messages, all chat messages, and private-dm messages
 
 ### Added
 
-- discordEmojis texture pack in assets folder
+- discordEmojis texture pack in a link
 - /toggleemojis command - turns emojis on/off
 - /discordemojis command - help command
 - All emojis you currently see in emojis.json